create board committe view, refactor code
